Move components to components folder to clean up
