+++ frontend/src/components/TagColorConfigEditor/TagColorConfigItemAdder.css
