+++ frontend/src/store/index.js
@@ -1,4 +1,4 @@
-import { createStore } from "redux";
+import { createStore } from 'redux';
 import rootReducer from "./reducers.js";
 
 const store = createStore(rootReducer);
